requirements.txt
argh==0.26.2
asn1crypto==0.24.0
attrs==18.2.0
autobahn==18.11.1
Automat==0.7.0
bitstring==3.1.5
cbor==1.0.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
constantly==15.1.0
crossbar==18.11.1
cryptography==2.3.1
h2==3.0.1
hpack==3.0.0
hyperframe==5.1.0
hyperlink==18.0.0
idna==2.5
incremental==17.5.0
Jinja2==2.10
lmdb==0.94
MarkupSafe==1.1.0
mistune==0.8.4
netaddr==0.7.19
numpy==1.15.4
opencv-python==3.4.3.18
passlib==1.7.1
pathtools==0.1.2
Pillow==5.3.0
pkg-resources==0.0.0
priority==1.3.0
psutil==5.4.8
py-ubjson==0.12.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
Pygments==2.2.0
PyHamcrest==1.9.0
PyNaCl==1.3.0
pyOpenSSL==18.0.0
PyQRCode==1.2.1
PyTrie==0.3.1
PyYAML==3.13
requests==2.20.1
sdnotify==0.3.2
service-identity==17.0.0
setproctitle==1.1.10
six==1.11.0
sortedcontainers==2.0.5
treq==18.6.0
Twisted==18.9.0
txaio==18.8.1
txtorcon==18.3.0
u-msgpack-python==2.5.0
urllib3==1.24.1
watchdog==0.9.0
zope.interface==4.6.0
